Moving response set items to their own folders
React components and redux containers get moved to a response_sets
folder
