+++ webpack/containers/FormsEditContainer.js
@@ -8,7 +8,7 @@ import { setSteps } from '../actions/tutorial_actions';
 import { fetchForm, saveForm, newForm, saveDraftForm } from '../actions/form_actions';
 import { addQuestion, removeQuestion, reorderQuestion, fetchQuestion } from '../actions/questions_actions';
 import FormEdit from '../components/FormEdit';
-import ResponseSetModal from './ResponseSetModal';
+import ResponseSetModal from './response_sets/ResponseSetModal';
 import QuestionModalContainer  from './QuestionModalContainer';
 import QuestionSearchContainer from './QuestionSearchContainer';
 import { formProps } from '../prop-types/form_props';
