+++ webpack/containers/QuestionModalContainer.js
@@ -11,8 +11,8 @@ import $ from 'jquery';
 import { saveQuestion } from '../actions/questions_actions';
 import Errors from '../components/Errors';
 import QuestionForm from '../components/QuestionForm';
-import ResponseSetList  from '../components/ResponseSetList';
-import ResponseSetDragWidget  from './ResponseSetDragWidget';
+import ResponseSetList  from '../components/response_sets/ResponseSetList';
+import ResponseSetDragWidget  from './response_sets/ResponseSetDragWidget';
 import { responseSetsProps }  from '../prop-types/response_set_props';
 import { fetchResponseTypes } from '../actions/response_type_actions';
 import { fetchQuestionTypes } from '../actions/question_type_actions';
