+++ webpack/containers/response_sets/ResponseSetEditContainer.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { setSteps } from '../actions/tutorial_actions';
-import { fetchResponseSet, saveResponseSet, saveDraftResponseSet } from '../actions/response_set_actions';
-import ResponseSetForm from '../components/ResponseSetForm';
-import { responseSetProps } from '../prop-types/response_set_props';
+import { setSteps } from '../../actions/tutorial_actions';
+import { fetchResponseSet, saveResponseSet, saveDraftResponseSet } from '../../actions/response_set_actions';
+import ResponseSetForm from '../../components/response_sets/ResponseSetForm';
+import { responseSetProps } from '../../prop-types/response_set_props';
 
 class ResponseSetEditContainer extends Component {
   constructor(props) {
